move deviceTools to core
