+++ src/core/deviceTools.spec.js
